Clarify and test for OS support - Ubuntu/Debian
AFAIK there isn't a `resolvconf` equivalent for EL, or any other non-Debian
distros. So raise an error if this module is used on such a system.

This might normally be done in a `::params` class. But we don't have enough
variables, especially vars shared across classes, to make it worthwhile. We
also only want to test for a single OS. So I've implemented it in the parent
class for the moment.