Flesh out example in README
I'm reluctant to give exhaustive details of the available options in the
README since they are well documented in the class using `puppetdoc` format.
Duplicating is likely to lead to one or the other being badly maintained.

I really wish there was a nice mechanism for rendiner `puppetdoc` to GitHub
or Forge READMEs.