[#3] Replace configs when they are symlinks
It turns out that `ensure => present, content => ''` on a path that is a
symlink will not do anything. `present` is satisfied and `content` is
ignored: http://projects.puppetlabs.com/issues/6808

This is a problem when some installations of `resolvconf` link
`resolv.conf.d/tail` to `resolv.conf.d/original`. We should always replace
this with the content we've defined.

/cc @maxamg