Pin dependencies in Gemfile
When running tests on Travis we remove the `Gemfile.lock` so that we
can install different versions of the `puppet` gem to test against.

Because there's no `Gemfile.lock` when Travis installs gems and we
previously didn't pin versions in the `Gemfile`, Travis will install
later versions of gems than the ones we want, which causes the tests
to fail.